kernel/mem/cache: replace `hold_last_slab` with `deallocate_last_available_slab`
`hold_last_slab` when true does not deallocate the last slab in the
entire cache, even if the other slabs are all in the full list meaning
there are no available slabs.

`deallocate_last_available_slab` when false does not deallocate the last
available slab when it becomes unused regardless to how many full slabs
there are, meaning there will always be an available slab.