kernel: add doccomment to `kernel_task` to clarify membership
The bootstrap init task is never in this list as it is never seen by the
scheduler nor is it allocated from the task cache so does not need to be
cleaned up.

Per-executor idle tasks are never in this list as they are handled
specially by the scheduler and are also not allocated from the task
cache and never clean up for the lifetime of the kernel.